Added Specification entitie and separated models and repositories inside modules
