+++ src/modules/cars/model/Category.ts
@@ -2,8 +2,11 @@ import { v4 as uuidV4 } from "uuid";
 
 class Category {
   id?: string;
+
   name: string;
+
   description: string;
+
   createdAt: Date;
 
   constructor() {
