+++ src/modules/repositories/CategoriesRepository.ts
@@ -1,4 +1,4 @@
-import { Category } from "../model/category";
+import { Category } from "../cars/model/Category";
 import {
   ICategoriesRepository,
   ICreateCategoryDTO,
