+++ src/modules/repositories/ICategoriesRepository.ts
@@ -1,4 +1,4 @@
-import { Category } from "../model/category";
+import { Category } from "../cars/model/Category";
 
 export interface ICreateCategoryDTO {
   name: string;
