+++ src/services/CreateSpecificationService.ts
@@ -1,0 +1,7 @@
+class CreateSpecificationService {
+  execute() {
+    console.log("TODO");
+  }
+}
+
+export { CreateSpecificationService };
